styling(admin):change nav color to blue
