+++ package.json
@@ -10,7 +10,6 @@
     "chart.js": "^2.8.0",
     "classnames": "^2.2.6",
     "crypto-random-string": "^3.0.1",
-    "email-existence": "^0.1.6",
     "jquery": "^3.4.1",
     "node-sass": "^4.12.0",
     "react": "^16.8.6",
